+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 
+## vNext
+- [Fix: Replaced non-threadsafe HashSet with ConcurrentDictionary in RequestTrackingTelemetryModule.IsHandlerToFilter](https://github.com/microsoft/ApplicationInsights-dotnet-server/pull/1211)
+
 ## Version 2.11.0-beta1
  - [Defer populating RequestTelemetry properties.](https://github.com/Microsoft/ApplicationInsights-dotnet-server/issues/1173)